requirements.txt
flask
kafka-python
pyodbc